proxy: move proxy package to internal
