+++ internal/proxy/mocks/app_conn_snapshot.go
